# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API para memorias de cálculo de instalaciones eléctricas según normativa NOM (México)",
        "title": "Garfex Calculadora de Filtros API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/calculos/conductor-tierra": {
            "post": {
                "description": "Selecciona el conductor de tierra según el ITM (interruptor termomagnético) y material",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Conductor"
                ],
                "summary": "Seleccionar conductor de tierra",
                "parameters": [
                    {
                        "description": "ITM y material del conductor",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.ConductorTierraRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Conductor de tierra seleccionado",
                        "schema": {
                            "$ref": "#/definitions/http.ConductorTierraResponse"
                        }
                    },
                    "400": {
                        "description": "Error de validación",
                        "schema": {
                            "$ref": "#/definitions/http.ConductorTierraResponseError"
                        }
                    },
                    "422": {
                        "description": "No se encontró conductor de tierra adecuado",
                        "schema": {
                            "$ref": "#/definitions/http.ConductorTierraResponseError"
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "$ref": "#/definitions/http.ConductorTierraResponseError"
                        }
                    }
                }
            }
        },
...
        "/calculos/memoria": {
            "post": {
                "description": "Calcula la memoria de cálculo completa para una instalación eléctrica según normativa NOM (México). Orquesta: corriente nominal → ajuste → conductor alimentación → conductor tierra → canalización → caída de tensión.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Memoria"
                ],
                "summary": "Memoria de cálculo completa",
                "parameters": [
                    {
                        "description": "Datos del equipo e instalación",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.CalcularMemoriaRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Memoria de cálculo exitosa",
                        "schema": {
                            "$ref": "#/definitions/http.CalcularMemoriaResponse"
                        }
                    },
                    "400": {
                        "description": "Error de validación o datos inválidos",
                        "schema": {
                            "$ref": "#/definitions/http.CalcularMemoriaResponseError"
                        }
                    },
                    "422": {
                        "description": "No se encontró conductor o canalización adecuada",
                        "schema": {
                            "$ref": "#/definitions/http.CalcularMemoriaResponseError"
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "$ref": "#/definitions/http.CalcularMemoriaResponseError"
                        }
                    }
                }
            }
        },
        "/calculos/tuberia": {
            "post": {
                "description": "Dimensiona el tamaño de tubería requerido según los calibres de conductores",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Canalización"
                ],
                "summary": "Dimensionar tubería",
                "parameters": [
                    {
                        "description": "Datos de conductores y tubería",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.CalcularTuberiaRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tamaño de tubería calculado",
                        "schema": {
                            "$ref": "#/definitions/http.CalcularTuberiaResponse"
                        }
                    },
                    "400": {
                        "description": "Error de validación",
                        "schema": {
                            "$ref": "#/definitions/http.CalcularTuberiaResponseError"
                        }
                    },
                    "422": {
                        "description": "Conductor no encontrado o tubería no disponible",
                        "schema": {
                            "$ref": "#/definitions/http.CalcularTuberiaResponseError"
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "$ref": "#/definitions/http.CalcularTuberiaResponseError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "dto.ConductorTierraOutput": {
            "type": "object",
            "properties": {
                "calibre": {
                    "description": "Calibre es el calibre del conductor seleccionado (ej: \"12 AWG\", \"2 AWG\").",
                    "type": "string"
                },
                "itm_hasta": {
                    "description": "ITMHasta es el ITM máximo para el cual aplica este calibre.",
                    "type": "integer"
                },
                "material": {
                    "description": "Material es el material del conductor (\"Cu\" o \"Al\").",
                    "type": "string"
                },
                "num_hilos": {
                    "description": "NumHilos es el número de hilos de tierra (1 para charola/tubería≤2, 2 para tubería\u003e2).",
                    "type": "integer"
                },
                "seccion_mm2": {
                    "description": "SeccionMM2 es la sección transversal en mm².",
                    "type": "number"
                }
            }
        },
...
        "dto.DatosEquipo": {
            "type": "object",
            "properties": {
                "amperaje": {
                    "description": "Qn/In - amperaje nominal o KVA/KVAR según tipo",
                    "type": "integer"
                },
                "bornes": {
                    "description": "Número de bornes (opcional)",
                    "type": "integer"
                },
                "clave": {
                    "description": "Clave comercial (ej: ACTISINE48D400MMB)",
                    "type": "string"
                },
                "itm": {
                    "description": "Interruptor termomagnético",
                    "type": "integer"
                },
                "tipo": {
                    "description": "\"A\", \"KVA\", \"KVAR\"",
                    "allOf": [
                        {
                            "$ref": "#/definitions/dto.TipoFiltro"
                        }
                    ]
                },
                "voltaje": {
                    "description": "Voltaje nominal (ej: 480)",
                    "type": "integer"
                }
            }
        },
...
        "dto.ModoCalculo": {
            "type": "string",
            "enum": [
                "LISTADO",
                "MANUAL_AMPERAJE",
                "MANUAL_POTENCIA"
            ],
            "x-enum-varnames": [
                "ModoListado",
                "ModoManualAmperaje",
                "ModoManualPotencia"
            ]
        },
...
        "dto.SistemaElectrico": {
            "type": "string",
            "enum": [
                "DELTA",
                "ESTRELLA",
                "BIFASICO",
                "MONOFASICO"
            ],
            "x-enum-varnames": [
                "SistemaElectricoDelta",
                "SistemaElectricoEstrella",
                "SistemaElectricoBifasico",
                "SistemaElectricoMonofasico"
            ]
        },
        "dto.TipoFiltro": {
            "type": "string",
            "enum": [
                "A",
                "KVA",
                "KVAR"
            ],
            "x-enum-comments": {
                "TipoFiltroA": "Filtro activo en Amperes",
                "TipoFiltroKVA": "Filtro calificado en KVA",
                "TipoFiltroKVAR": "Filtro de rechazo en KVAR"
            },
            "x-enum-varnames": [
                "TipoFiltroA",
                "TipoFiltroKVA",
                "TipoFiltroKVAR"
            ]
        },
...
        "http.CalcularMemoriaRequest": {
            "type": "object",
            "required": [
                "estado",
                "longitud_circuito",
                "modo",
                "sistema_electrico",
                "tension",
                "tipo_canalizacion",
                "tipo_voltaje"
            ],
            "properties": {
                "amperaje_nominal": {
                    "type": "number"
                },
                "diametro_control_mm": {
                    "type": "number"
                },
                "equipo": {
                    "description": "Equipo completo (modo LISTADO)",
                    "allOf": [
                        {
                            "$ref": "#/definitions/dto.DatosEquipo"
                        }
                    ]
                },
                "estado": {
                    "type": "string"
                },
                "factor_potencia": {
                    "type": "number"
                },
                "hilos_por_fase": {
                    "type": "integer"
                },
                "itm": {
                    "description": "Requerido en MANUAL_*, opcional en LISTADO (usa equipo.itm)",
                    "type": "integer"
                },
                "longitud_circuito": {
                    "type": "number"
                },
                "material": {
                    "description": "material: Cu (cobre), Al (aluminio)",
                    "type": "string"
                },
                "modo": {
                    "description": "Modo de cálculo (required)\nValues: LISTADO, MANUAL_AMPERAJE, MANUAL_POTENCIA",
                    "allOf": [
                        {
                            "$ref": "#/definitions/dto.ModoCalculo"
                        }
                    ]
                },
                "num_tuberias": {
                    "type": "integer"
                },
                "porcentaje_caida_maximo": {
                    "type": "number"
                },
                "potencia_nominal": {
                    "type": "number"
                },
                "potencia_unidad": {
                    "description": "potencia_unidad: W, KW, KVA, KVAR",
                    "type": "string"
                },
                "sistema_electrico": {
                    "description": "Sistema eléctrico\nValues: DELTA, ESTRELLA, BIFASICO, MONOFASICO",
                    "allOf": [
                        {
                            "$ref": "#/definitions/dto.SistemaElectrico"
                        }
                    ]
                },
                "temperatura_override": {
                    "type": "integer"
                },
                "tension": {
                    "description": "═══════════════════════════════════════════════════════════════════════\nDATOS DE INSTALACIÓN (comunes a todos los modos)\n═══════════════════════════════════════════════════════════════════════",
                    "type": "number"
                },
                "tension_unidad": {
                    "type": "string"
                },
                "tipo_canalizacion": {
                    "description": "tipo_canalizacion: TUBERIA_PVC, TUBERIA_ALUMINIO, TUBERIA_ACERO_PG, TUBERIA_ACERO_PD, CHAROLA_CABLE_ESPACIADO, CHAROLA_CABLE_TRIANGULAR",
                    "type": "string"
                },
                "tipo_equipo": {
                    "description": "Datos manuales (modo MANUAL_*)\ntipo_equipo: FILTRO_ACTIVO, TRANSFORMADOR, FILTRO_RECHAZO, CARGA",
                    "type": "string"
                },
                "tipo_voltaje": {
                    "description": "tipo_voltaje: FASE_NEUTRO, FASE_FASE",
                    "type": "string"
                }
            }
        },
        "http.CalcularMemoriaResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/dto.MemoriaOutput"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "http.CalcularMemoriaResponseError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "http.CalcularTuberiaRequest": {
            "type": "object",
            "required": [
                "calibre_fase",
                "calibre_tierra",
                "num_fases",
                "num_tuberias",
                "tipo_canalizacion"
            ],
            "properties": {
                "calibre_fase": {
                    "type": "string"
                },
                "calibre_neutral": {
                    "type": "string"
                },
                "calibre_tierra": {
                    "type": "string"
                },
                "num_fases": {
                    "type": "integer"
                },
                "num_neutros": {
                    "type": "integer",
                    "minimum": 0
                },
                "num_tuberias": {
                    "type": "integer"
                },
                "tipo_canalizacion": {
                    "type": "string"
                }
            }
        },
        "http.CalcularTuberiaResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/dto.TuberiaOutput"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "http.CalcularTuberiaResponseError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "http.ConductorTierraRequest": {
            "type": "object",
            "required": [
                "itm"
            ],
            "properties": {
                "itm": {
                    "description": "ITM es el Interruptor Termomagnético en amperes (requerido, \u003e 0).",
                    "type": "integer"
                },
                "material": {
                    "description": "Material es el material del conductor (\"Cu\" o \"Al\").\nSi está vacío, se usa \"Cu\" por defecto.",
                    "type": "string"
                }
            }
        },
        "http.ConductorTierraResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/dto.ConductorTierraOutput"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "http.ConductorTierraResponseError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
    }
}